requirements.txt
django-model-utils >= 4.1.1
# Additional requirements go here
channels==3.0.3
